requirements.txt
Flask
Flask-Migrate
Flask-SQLAlchemy
Flask-cors
mysql-connector
mysqlclient
requests
pymysql
